+++ datanode/entities/referral_set_stats.go
@@ -29,6 +29,7 @@ type (
 		AtEpoch                               uint64
 		WasEligible                           bool
 		ReferralSetRunningNotionalTakerVolume string
+		ReferrerTakerVolume                   string
 		RefereesStats                         []*eventspb.RefereeStats
 		VegaTime                              time.Time
 		RewardFactor                          string
@@ -41,6 +42,7 @@ type (
 		AtEpoch                               uint64
 		WasEligible                           bool
 		ReferralSetRunningNotionalTakerVolume string
+		ReferrerTakerVolume                   string
 		VegaTime                              time.Time
 		PartyID                               string
 		DiscountFactor                        string
@@ -78,6 +80,7 @@ func (s FlattenReferralSetStats) ToProto() *v2.ReferralSetStats {
 	return &v2.ReferralSetStats{
 		AtEpoch:                               s.AtEpoch,
 		ReferralSetRunningNotionalTakerVolume: s.ReferralSetRunningNotionalTakerVolume,
+		ReferrerTakerVolume:                   s.ReferrerTakerVolume,
 		PartyId:                               s.PartyID,
 		DiscountFactor:                        s.DiscountFactor,
 		RewardFactor:                          s.RewardFactor,
@@ -105,6 +108,7 @@ func ReferralSetStatsFromProto(proto *eventspb.ReferralSetStatsUpdated, vegaTime
 		AtEpoch:                               proto.AtEpoch,
 		WasEligible:                           proto.WasEligible,
 		ReferralSetRunningNotionalTakerVolume: proto.ReferralSetRunningNotionalTakerVolume,
+		ReferrerTakerVolume:                   proto.ReferrerTakerVolume,
 		RefereesStats:                         proto.RefereesStats,
 		VegaTime:                              vegaTime,
 		RewardFactor:                          proto.RewardFactor,