+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [9408](https://github.com/vegaprotocol/vega/issues/9408) - Enforce pagination range.
 - [9757](https://github.com/vegaprotocol/vega/issues/9757) - Liquidity provisions `API` shows the pending `LP` instead of the current when an update is accepted by the network.
 
+
 ### 🗑️ Deprecation
 
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -
@@ -176,6 +177,7 @@
 - [9770](https://github.com/vegaprotocol/vega/issues/9770) - Fix `PnL` flickering bug on an already open position.
 - [9827](https://github.com/vegaprotocol/vega/issues/9827) - Expose `was_eligible` in the referral set statistics.
 - [9918](https://github.com/vegaprotocol/vega/issues/9918) - Expose vesting quantum balance.
+- [9923](https://github.com/vegaprotocol/vega/issues/9923) - Add referrer volume to referral set stats.
 
 ### 🐛 Fixes